+++ src/main/java/com/gpxparser/SrtFileReaderLambda.java
@@ -8,10 +8,14 @@ import java.io.BufferedReader;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.Month;
 import java.time.ZoneId;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -67,10 +71,10 @@ public class SrtFileReaderLambda extends SrtFileReader {
                 // extract and build GPS longitude, latitude and elevation data
                 matcher = gpsLinePattern.matcher(line);
                 if (matcher.matches()) {
-                    Double longitude = Double.parseDouble(matcher.group(1));
-                    Double latitude = Double.parseDouble(matcher.group(2));
-                    Double gpsAltitude = 10 * Double.parseDouble(matcher.group(3));
-                    Double elevation = Double.parseDouble(matcher.group(4));
+                    BigDecimal longitude = BigDecimal.valueOf(Double.parseDouble(matcher.group(1)));
+                    BigDecimal latitude = BigDecimal.valueOf(Double.parseDouble(matcher.group(2)));
+                    BigDecimal gpsAltitude = BigDecimal.valueOf(10 * Double.parseDouble(matcher.group(3)));
+                    BigDecimal elevation = BigDecimal.valueOf(Double.parseDouble(matcher.group(4)));
 
                     srtBlock = srtPointList.getLast();
                     srtBlock.setLatitude(latitude);
